mechanic mon

+++ Assets/0Game/Script/MainSystem/CalculatorDamageSystem.cs
@@ -74,7 +74,7 @@ public class CalculatorDamageSystem : MonoBehaviour
         return total;
     }
 
-    public int DamageResult(int atk, int skillDMG, int targetDef, int damageBonus = 0, int targetDMGRe = 0)
+    public int DamageResult(float atk, int skillDMG, int targetDef, float damageBonus = 0, float targetDMGRe = 0)
     {
         int damage = (int)Mathf.Floor((((atk * skillDMG / 100) * ((100 + damageBonus) / 100)) - targetDef) * (100 - targetDMGRe) / 100);
         //if (damage <= 0) damage = 1;
