+++ Assets/0Game/Script/MainSystem/DamageCalculator.cs
@@ -12,7 +12,7 @@ public class DamageCalculator : MonoBehaviour
         
     }
 
-    public int DamageResult(int atk, int skillDMG, int targetDef, int damageBonus = 0, int targetDMGRe = 0)
+    public int DamageResult(float atk, int skillDMG, int targetDef, float damageBonus = 0, float targetDMGRe = 0)
     {
         int damage = (int)Mathf.Floor( (( (atk * skillDMG / 100) * (100 + damageBonus)/100) - targetDef) * (100 - targetDMGRe) / 100);
         if (damage <= 0) damage = 1;
